move reader.py from edinet to xbrl
