+++ tests/edinet/reader/aspects/test_business.py
@@ -1,6 +1,6 @@
 import os
 import unittest
-from xbrr.edinet.reader.reader import Reader
+from xbrr.xbrl.reader.reader import Reader
 from xbrr.edinet.reader.aspects.business import Business
 import tests.edinet.reader.doc as testdoc
