+++ tests/edinet/test_api.py
@@ -22,7 +22,7 @@ class TestAPI(unittest.TestCase):
 
     @delay
     def test_api_documents(self):
-        documents = xbrr.edinet.api.documents.get("2019-02-01")
+        documents = xbrr.edinet.api.documents.get("2022-02-01")
         self.assertEqual(documents.metadata.count, len(documents.list))
 
         _dir = os.path.dirname(__file__)
