+++ xbrr/edinet/facade.py
@@ -1,7 +1,7 @@
 from xbrr.edinet.client.document_list_client import MetaDataClient
 from xbrr.edinet.client.document_list_client import DocumentListClient
 from xbrr.edinet.client.document_client import DocumentClient
-from xbrr.edinet.reader.reader import Reader
+from xbrr.xbrl.reader.reader import Reader
 from xbrr.edinet.reader.aspects.business import Business
 from xbrr.edinet.reader.aspects.company import Company
 from xbrr.edinet.reader.aspects.finance import Finance
