+++ xbrr/edinet/reader/aspects/finance.py
@@ -5,7 +5,7 @@ import importlib
 if importlib.util.find_spec("pandas") is not None:
     import pandas as pd
 from xbrr.base.reader.base_parser import BaseParser
-from xbrr.edinet.reader.element_value import ElementValue
+from xbrr.xbrl.reader.element_value import ElementValue
 
 
 class Finance(BaseParser):
