+++ xbrr/edinet/reader/aspects/metadata.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from xbrr.base.reader.base_parser import BaseParser
-from xbrr.edinet.reader.element_value import ElementValue
+from xbrr.xbrl.reader.element_value import ElementValue
 
 
 class Metadata(BaseParser):
