+++ xbrr/edinet/reader/aspects/stock.py
@@ -1,5 +1,5 @@
 from xbrr.base.reader.base_parser import BaseParser
-from xbrr.edinet.reader.element_value import ElementValue
+from xbrr.xbrl.reader.element_value import ElementValue
 
 
 class Stock(BaseParser):
