+++ xbrr/xbrl/reader/element_schema.py
