+++ xbrr/xbrl/reader/role_schema.py
